Extract ChannelConfig from HttpClientConfig
